requirements.txt
itsdangerous
Jinja2
MarkupSafe
wheel
gunicorn
Flask-OAuthlib
decorator
Flask
Flask-SQLAlchemy
Flask-WTF
lxml
pbr
Pillow
PyGithub
six
SQLAlchemy
sqlalchemy-migrate
sqlparse
Tempita
Werkzeug
WTForms
Flask-pymongo
dnspython
flask_socketio
passlib
